+++ tests/components/supernotify/hass_setup_lib.py
@@ -49,8 +49,10 @@ from custom_components.supernotify.archive import NotificationArchive
 from custom_components.supernotify.common import DupeChecker
 from custom_components.supernotify.context import Context
 from custom_components.supernotify.delivery import Delivery, DeliveryRegistry
+from custom_components.supernotify.envelope import Envelope
 from custom_components.supernotify.hass_api import HomeAssistantAPI
 from custom_components.supernotify.media_grab import MediaStorage
+from custom_components.supernotify.notification import Notification
 from custom_components.supernotify.notify import TRANSPORTS
 from custom_components.supernotify.people import PeopleRegistry
 from custom_components.supernotify.scenario import ScenarioRegistry
@@ -74,6 +76,10 @@ def load_config(v: str | dict | list | None, return_type: type = dict) -> JSON_T
     if not v:
         return return_type()
     return v
+
+
+def first_envelope(notification: Notification, delivery: str) -> Envelope:
+    return notification.deliveries[delivery]["delivered"][0]  # type: ignore
 
 
 class TestingContext(Context):